passes all tests

+++ src/decode.cuh
@@ -12,18 +12,18 @@ __global__ void applyErrors(int* lookup, curandState_t* states, int* errorTarget
 
 __global__ void calculateSyndrome(int* lookup, int* qubits, int* syndrome, int* edgeToFaces);
 
-__global__ void flip(int* qLookup, int* sLookup, int* qubits, int* syndrome, int* faceToEdges, int* edgeToFaces, int* qubitMessages, int* qubitMarginals);
+__global__ void flip(int* qLookup, int* sLookup, int* qubits, int* syndrome, int* faceToEdges, int* edgeToFaces, float* qubitMessages, float* qubitMarginals);
 
 __global__ void pflip(int* qLookup, int* sLookup, int* qubits, int* syndrome, int* faceToEdges, 
-                        int* edgeToFaces, int* qubitMessages, int* qubitMarginals, curandState_t* states);
+                        int* edgeToFaces, float* qubitMessages, float* qubitMarginals, curandState_t* states);
 
-__global__ void updateSyndromeMessages(int* lookup, int* qubitMessages, int* syndrome, int* syndromeMessages, int* edgeToFaces, int* faceToEdges);
+__global__ void updateSyndromeMessages(int* lookup, float* qubitMessages, int* syndrome, float* syndromeMessages, int* edgeToFaces, int* faceToEdges);
 
-__global__ void updateQubitMessages(int* lookup, int* qubitMessages, int* syndromeMessages, int* faceToEdges, int* edgeToFaces, int p);
+__global__ void updateQubitMessages(int* lookup, float* qubitMessages, float* syndromeMessages, int* faceToEdges, int* edgeToFaces, float p);
 
-__global__ void calcMarginals(int* lookup, int* qubitMarginals, int* syndromeMessages, int p);
+__global__ void calcMarginals(int* lookup, float* qubitMarginals, float* syndromeMessages, float p);
 
-__global__ void bpCorrection(int* lookup, int* qubits, int* qubitMarginals);
+__global__ void bpCorrection(int* lookup, int* qubits, float* qubitMarginals);
 
 __global__ void measureLogicals(int* lookup, int* qubits, int* nOdd, int L, char bounds);
